pyproject.toml
[tool.poetry]
name = "reviro"
version = "0.1.0"
description = ""
authors = ["Anas Kaarov <anasbinkurban@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.1.6"
django-4-jet = "^1.0.9"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
django-filter = "^22.1"
drf-spectacular = "^0.27.1"
django-rest-swagger = "^2.2.0"
gunicorn = "^20.1.0"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"
factory-boy = "^3.2.1"
pytest = "^7.2.2"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
qrcode = "^7.4.2"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
